refactor(supervision): consolidate mission watchdogs into api/supervision.rs (Phase 5)
recover_server_shutdown_missions, stuck_mission_watchdog_loop (incl.
OOM-kill scope checks), ack_promotion_loop, and stale_mission_cleanup_loop
move verbatim out of control.rs into one supervision module, so the three
overlapping liveness mechanisms finally live side by side. Module header
documents the Phase 5b follow-up: a single per-mission LivenessState to
replace their independent 'is this mission alive?' heuristics (the dual
'stalled' notions caused past watchdog false-positives). No behavior
change.
